requirements.txt
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.3
pillow==10.0.1
urllib3==2.1.0
gunicorn==21.2.0
